Move the positivity checker to its own folder
