+++ src/Juvix/Syntax/MicroJuvix/Language/Extra.hs
@@ -389,14 +389,6 @@ unfoldExpressionApp = \case
 unfoldApplication :: Application -> (Expression, NonEmpty Expression)
 unfoldApplication = fmap (fmap snd) . unfoldApplication'
 
-nameInExpression :: Name -> Expression -> Bool
-nameInExpression n = \case
-  ExpressionIden (IdenVar var) -> n == var
-  ExpressionIden (IdenInductive ty) -> n == ty
-  ExpressionApplication (Application l r _) -> nameInExpression n l || nameInExpression n r
-  ExpressionFunction (Function l r) -> nameInExpression n (l ^. paramType) || nameInExpression n r
-  _ -> False
-
 reachableModules :: Module -> [Module]
 reachableModules = fst . run . runOutputList . evalState (mempty :: HashSet Name) . go
   where
